Added AI, cant clean up old targeting system though...

+++ update.h
@@ -1,7 +1,7 @@
 #ifndef _UPDATE_
 #define _UPDATE_
 
-void Move_Ent(Entity *thisEnt, int xAmnt, int yAmnt);
+void Move_Ent(Entity *thisEnt, double xAmnt, double yAmnt);
 void update();
 void DrawUI();
 void LevelTower(int towerNum);
